week04-homework/pyproject.toml
[project]
name = "week04-homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"langchain>=1.2.8",
	"langchain-core>=1.2.9",
	"langchain-deepseek>=1.0.1",
	"langgraph>=1.0.7",
	"python-dotenv",
	# add your dependencies here and run `uv sync` to install them
]
